refactor: move With to own file
