+++ riscv_analysis/src/parser/with_has_raw_text.rs
@@ -1,0 +1,7 @@
+use super::{HasRawText, With};
+
+impl<T> HasRawText for With<T> {
+    fn raw_text(&self) -> &str {
+        self.text.as_str()
+    }
+}
